More design and planning
Describe how a package is added and how the BinPackage class should
look like.

Summarize the relevant passages of the Debian Repository Format
and maybe from some more places in a new document standards.rst where
we also move the source snippets to.

Finally implement parallel checksumming for MD5, SHA1 and SHA256 in
a utility.