Base BinPackage on SimpleNamespace again and fix pool functions
Our BinPackage based classes are not adequetely represented by named tuples
because these are immutable. Using a mix of mutable and immutable properties
is akward, so go back to SimpleNamespace. The necessary changes are
surprisingly small.

Actually add the package to the store now, which now also sets the 'Filename'
attribute, that we need to set in the database as well. The __str__
method of BinPackageDb now shows it as well.

Fixes in pool:
- Fix pure_version: Compare sep to '', not None
- Use os.path.join to compose paths
- Make sure the target directory in a pool exists