Implement full package add algorithm
First we slightly redesign how references are returned from the db
module: Return a list, with the target release entry guaranteed to be
the first. Also return a list of BinPkgRef objects and also add the
stored filename, as the store will need it.

Now add (pseudo) release and comoonent args to addBinary in main
and add the rather longwinded algorithm, and remove the comments
where it is outlined (incompletely, by the way)

Fix an SQL syntax error in Db.replaceBinary, add a delBinaryRef method.

A new package obtained from file gets id -1 initially

A few TODO items added, approximately in priority order