Start implementation of indexing
There is a new db method getIndex which returns a package generator
that iterates through all packages of a given arch, component of a
release.

The new BinIndexer class openes the package file and writes package
by package

The main program is changed to obtain the action as first argument.