Rewamp config handling
Config is now an object returned by the factory function config.getConfig.
The conversion of a dict obtained from yaml to an object is then done
as soon as possible and almost all operations on the config are written
in object notation. To prevent arbitrary attributes to be set by a
strange config, we clean up the dictionary using the _top_options
dict which also supplies static default values.
The same procedure is followed when creating Release objects.

More small changes to config.py
- add a logger
- defarchitectures is a set now, release.architectures should be a set
  as well
- gpgkey in to renamed to defgpgkey

In sqlite, during initialization, we not only create new releases from
the config, we also set the ids of releases present in the config,
the new method is called syncreleases and replaces th former
mkreleases and makes relId superflous. We now warn when a release found
in the database is not mentioned in the config.

In the main program, we have a new getRelease function that obtains
the release objet from the name and warns when an unknown release
is encountered. The former variable 'release' is now replaced by
relname, and release is the Release object. And of course the config
is created by the new factory now.