Implement Release creation and signing
The close method of a CsumCompressor now creates a list of (filename, Hash)
tuples instead of a new merged tuple. BinIndexer then returns that list of
tuples instead of printing it.

_updateRelease steps through all components and architectures, updates
the indices, collects checksum information and finally writes everything
into the Release file. _signRelease adds proper signatures.

The main program now calls updateRelease with the release passed on
the commandline. It also sets the time locale to C to get English
month- and day-of-week names.

The getIndex function in sqlite gets a new boolean keyword argument
'with_all', to either include the 'all' architecture together with a
binary one or not. Default is not.

Finally add some minor info to our standards document.