Introduce componentrules
The special commandline arguments R= and S= don't seem to be such a good
idea after alll, and it is much easier to just issue different debrep
commands instead of building a large commandline by hand.

On the other hand, offering a way to automatically determine the component
does seem like a good idea, so go for the simples way and that is to allow
the config to specify a list of rules that is evaluated in order.
Each rule is a list of glob patterns and a component name. If any of the
pattenrms matches, the current component is use. If no match is found, use
the default component as before.

- Implement a new config method getPkgComponent that does the actual
  work
- Change determining the component in doAdd. Note that we now need to
  do getBinFromDeb in doAdd, because we need the package's name. And no,
  we can't take it from the file name, because we allow arbitrary
  filenames for our package files.
- Update the manpage