Implement deletes
As the SQL queries to select packages for listing and for deletion
have similar - but not quite the same - structure, we move the creation
of the query, together with argument generation to some common utility
functions.

The deletion function in the db module returns a result which either is
a list of still existing references to the package or the essential data
of the package itself.

In debrep we move the getRelease function to doAdd, the only place it is
called, but create a relidsFromArgs that can be called from doList and doDel,
with a sligth variation in what an unspecified release means: for doList, it
means: list all releases, whereas in doDel it means: only delete from the
default release.

str2list is moved into the general helper function arg2list