Store package Filename in the reference
Only the debian package pool store uses a Filename key in the Package
index that is independent of release and component, so as we want to
support different names for the same actual package, the Filename needs
to be stored in the reference. The same issue will arise for source
packages, so for those we also store the Directory in the reference.

Also the store functions that add a package or reference now return
the filename and the db functions take a filename

Make the source name of a package a general BinPackage function.
Same for poolDir although that doesn't look so useful for other stores.

Also, the config now caches the actual store and db internally, so the
getDb and getStore might be called several times. The actual reason is
that we want to give the store access to the database.