Fix package deletion
As the Filename of a package is now stored in the reference, we can't just
delete the reference first. Instead fetch all references prior to a delete
and later return that list, possibly with idref set to None to mark the
deletion of the last package.

Also remove the check in getrefsAdd that only a single version of a package
may be contained in a release, as that will become optional and the check
doesn't belong there anyway.