More fixes for Deletion
Chnage the name and return value of the db deletion function again.
Note that for non pools we must return all current references with all
information there complete, so just add a 'deleted' property and for
easier store handling move the deleted ref to the front.

Also fix removal of the last ref and add proper debugging messages