Config fixes
When the config is read warn about unknown items and delete them.
Also fix the deletion code to not modify the dict that we are
iterating over.

Also search the debrep config file in the current directory, not in
the subdirectory 'config' of the current directory which is quite
unintuitive and uncommon. (Documentation was already adapted in the
last commit)