Move common store functions to base class
A store is now a subclass of BaseStore which provides common code.

The methods used for an abstract store and the methods provided
by base classes are now documented in the new basestore.py source
file.

The binNewPkg and binAddRef methods do not return the newly
computed Filename, they just require that the Filename be set in
the passed pkg object.

The database argument was removed from the Store constructor as it
is obtained from the passed config object

The poolDir method was removed from the binPackage class as it
is only used in the pool store. In the pool store it is no separate
function any more, but made part of pkgDir

A common binDelCleanup method has been added to the store base class
that removes a File and all now empty directories is was in.